Break out StroyboardNavInitial
